Actualización de librerías Acomodar proyectos de prueba
