+++ Cashflowio.sln
@@ -13,7 +13,9 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Cashflowio.Infrastructure",
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Cashflowio.Web", "src\Cashflowio.Web\Cashflowio.Web.csproj", "{C9751CB7-4CD6-4633-A99A-4F6ADD525437}"
 EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Cashflowio.Tests", "tests\Cashflowio.Tests\Cashflowio.Tests.csproj", "{AEE17BAB-D187-4BC5-B640-40CC0749566C}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Cashflowio.IntegrationTests", "tests\Cashflowio.IntegrationTests\Cashflowio.IntegrationTests.csproj", "{AEE17BAB-D187-4BC5-B640-40CC0749566C}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Cashflowio.UnitTests", "tests\Cashflowio.UnitTests\Cashflowio.UnitTests.csproj", "{9613ED92-A5C9-463F-82DB-087FF1595FF6}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -73,6 +75,18 @@ Global
 		{AEE17BAB-D187-4BC5-B640-40CC0749566C}.Release|x64.Build.0 = Release|Any CPU
 		{AEE17BAB-D187-4BC5-B640-40CC0749566C}.Release|x86.ActiveCfg = Release|Any CPU
 		{AEE17BAB-D187-4BC5-B640-40CC0749566C}.Release|x86.Build.0 = Release|Any CPU
+		{9613ED92-A5C9-463F-82DB-087FF1595FF6}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{9613ED92-A5C9-463F-82DB-087FF1595FF6}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{9613ED92-A5C9-463F-82DB-087FF1595FF6}.Debug|x64.ActiveCfg = Debug|Any CPU
+		{9613ED92-A5C9-463F-82DB-087FF1595FF6}.Debug|x64.Build.0 = Debug|Any CPU
+		{9613ED92-A5C9-463F-82DB-087FF1595FF6}.Debug|x86.ActiveCfg = Debug|Any CPU
+		{9613ED92-A5C9-463F-82DB-087FF1595FF6}.Debug|x86.Build.0 = Debug|Any CPU
+		{9613ED92-A5C9-463F-82DB-087FF1595FF6}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{9613ED92-A5C9-463F-82DB-087FF1595FF6}.Release|Any CPU.Build.0 = Release|Any CPU
+		{9613ED92-A5C9-463F-82DB-087FF1595FF6}.Release|x64.ActiveCfg = Release|Any CPU
+		{9613ED92-A5C9-463F-82DB-087FF1595FF6}.Release|x64.Build.0 = Release|Any CPU
+		{9613ED92-A5C9-463F-82DB-087FF1595FF6}.Release|x86.ActiveCfg = Release|Any CPU
+		{9613ED92-A5C9-463F-82DB-087FF1595FF6}.Release|x86.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
@@ -82,6 +96,7 @@ Global
 		{220361D6-9C76-4E3F-BD34-3C7B50E2CA4D} = {106AE906-5075-410A-B941-912F811848EE}
 		{C9751CB7-4CD6-4633-A99A-4F6ADD525437} = {106AE906-5075-410A-B941-912F811848EE}
 		{AEE17BAB-D187-4BC5-B640-40CC0749566C} = {B31B4797-1D9F-4288-808C-BE9A31A98C7D}
+		{9613ED92-A5C9-463F-82DB-087FF1595FF6} = {B31B4797-1D9F-4288-808C-BE9A31A98C7D}
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
 		SolutionGuid = {B0F19343-8185-4A9F-9165-0EA8544BC925}
