+++ tests/Cashflowio.IntegrationTests/Assets/CoinKeeper.xlsx
